Refector: Cart, Order, Payment Layer Exception

+++ src/main/java/com/example/demo/cart/controller/CartController.java
@@ -17,6 +17,7 @@ import org.springframework.web.bind.annotation.RestController;
 
 import com.example.demo.cart.dto.CartItemAddReq;
 import com.example.demo.cart.dto.CartRes;
+import com.example.demo.cart.exception.UnauthorizedException;
 import com.example.demo.cart.service.CartService;
 import com.example.demo.user.entity.UserEntity;
 
@@ -33,12 +34,18 @@ public class CartController {
 	@PostMapping("/add")
 	public ResponseEntity<String> addItems(@RequestBody @Valid CartItemAddReq req,
 		@AuthenticationPrincipal UserEntity user) {
+		if (user == null) {
+			throw new UnauthorizedException();
+		}
 		cartService.addItemToCart(req, user.getUserId());
 		return ResponseEntity.ok("장바구니에 추가되었습니다.");
 	}
 
 	@GetMapping
 	public ResponseEntity<List<CartRes>> getMyCart(@AuthenticationPrincipal UserEntity user) {
+		if (user == null) {
+			throw new UnauthorizedException();
+		}
 		List<CartRes> myCart = cartService.getMyCart(user.getUserId());
 		return ResponseEntity.ok(myCart);
 	}
@@ -46,12 +53,18 @@ public class CartController {
 	@PatchMapping("/items/{cartItemId}")
 	public ResponseEntity<String> updateQuantity(@PathVariable UUID cartItemId, @RequestParam int quantity,
 		@AuthenticationPrincipal UserEntity user) {
+		if (user == null) {
+			throw new UnauthorizedException();
+		}
 		cartService.updateQuantity(cartItemId, quantity, user.getUserId());
 		return ResponseEntity.ok("수량이 변경되었습니다.");
 	}
 
 	@DeleteMapping("/items/{cartItemId}")
 	public ResponseEntity<String> deleteItem(@PathVariable UUID cartItemId, @AuthenticationPrincipal UserEntity user) {
+		if (user == null) {
+			throw new UnauthorizedException();
+		}
 		cartService.deleteItem(cartItemId, user.getUserId());
 		return ResponseEntity.ok("장바구니 항목이 삭제되었습니다.");
 	}
